More code generation fixes

+++ runtime/src/test/java/org/qi4j/test/ASMTest.java
@@ -82,7 +82,7 @@ public void createClass()
             {
             }
 
-            public int unwrapResult()
+            public long unwrapResult()
             {
                 return 0;
             }
